+++ src/main/java/com/adit/backend/domain/user/service/command/FriendCommandService.java
@@ -7,6 +7,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.adit.backend.domain.notification.service.NotificationGenerationService;
+import com.adit.backend.domain.notification.service.command.NotificationCommandService;
 import com.adit.backend.domain.user.converter.FriendConverter;
 import com.adit.backend.domain.user.dto.request.FriendRequestDto;
 import com.adit.backend.domain.user.dto.response.FriendshipResponseDto;
@@ -29,12 +31,16 @@ public class FriendCommandService {
 
 	private final UserRepository userRepository;
 	private final FriendshipRepository friendshipRepository;
+	private final NotificationCommandService notificationCommandService;
 	private final FriendConverter friendConverter;
+	private final NotificationGenerationService notificationGenerationService;
 
 	// 친구 요청 보내기
 	public FriendshipResponseDto sendFriendRequest(FriendRequestDto requestDto) {
-		User FromUser = userRepository.findById(requestDto.fromUserId()).orElseThrow(() -> new UserException(USER_NOT_FOUND));
-		User toUser = userRepository.findById(requestDto.toUserId()).orElseThrow(() -> new UserException(USER_NOT_FOUND));
+		User FromUser = userRepository.findById(requestDto.fromUserId())
+			.orElseThrow(() -> new UserException(USER_NOT_FOUND));
+		User toUser = userRepository.findById(requestDto.toUserId())
+			.orElseThrow(() -> new UserException(USER_NOT_FOUND));
 		// 친구 요청(정방향)
 		Friendship forwardRequest = friendConverter.toEntity(FromUser, toUser, true);
 		// 친구 요청(역방향)
@@ -42,7 +48,7 @@ public class FriendCommandService {
 
 		Friendship savedForwardRequest = friendshipRepository.save(forwardRequest);
 		friendshipRepository.save(reverseRequest);
-
+		notificationGenerationService.createNotificationOfAFriendRequest(savedForwardRequest);
 		return friendConverter.toResponse(savedForwardRequest);
 	}
 
@@ -50,8 +56,8 @@ public class FriendCommandService {
 	public void acceptFriendRequest(Long requestId) {
 		Friendship friendRequest = friendshipRepository.findById(requestId)
 			.orElseThrow(() -> new FriendShipException(FRIEND_REQUEST_NOT_FOUND));
-
 		friendRequest.acceptRequest();
+		notificationGenerationService.createNotificationOfAFriendAccept(friendRequest);
 	}
 
 	// 친구 요청 거절
